requirements.txt
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mysqlclient==2.0.3
numpy==1.19.5
pandas==1.2.1
patsy==0.5.1
Pillow==8.1.0
progressbar2==3.53.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2020.5
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
statsmodels==0.12.2
threadpoolctl==2.1.0
Werkzeug==1.0.1
